refactor: ref all component in homepage
