+++ src/components/common/AnimatedSection.tsx
@@ -1,0 +1,45 @@
+import { motion, type Variants } from 'framer-motion';
+import React from 'react';
+
+interface AnimatedSectionProps {
+  children: React.ReactNode;
+  className?: string;
+}
+
+const containerVariants: Variants = {
+  hidden: { opacity: 0 },
+  visible: {
+    opacity: 1,
+    transition: {
+      staggerChildren: 0.2,
+    },
+  },
+};
+
+const itemVariants: Variants = {
+  hidden: { y: 20, opacity: 0 },
+  visible: {
+    y: 0,
+    opacity: 1,
+    transition: {
+      duration: 0.5,
+      ease: 'easeOut',
+    },
+  },
+};
+
+const AnimatedSection: React.FC<AnimatedSectionProps> = ({ children, className = '' }) => {
+  return (
+    <motion.div
+      className={className}
+      variants={containerVariants}
+      initial="hidden"
+      whileInView="visible"
+      viewport={{ amount: 0.4 }}
+    >
+      {children}
+    </motion.div>
+  );
+};
+
+export { AnimatedSection, itemVariants };
