+++ src/components/common/index.ts
@@ -1,0 +1,4 @@
+export { default as AnimatedText } from './AnimatedText';
+export { default as SkillCardComponent } from './SkillCardComponent';
+export { default as ProjectCard } from './ProjectCard';
+export { AnimatedSection, itemVariants } from './AnimatedSection';
