+++ src/components/home/index.ts
@@ -1,0 +1,2 @@
+export { AnimatedHero } from './HeroSection';
+export { AnimatedCTA } from './CTASection';
