+++ src/pages/index.astro
@@ -20,7 +20,6 @@ import { portofolioItems } from '@/lib/allData';
         darkGridColor="#374151"
         client:visible
     >
-        <!-- Hapus HTML lama dan panggil komponen Hero animasi -->
         <AnimatedHero client:visible />
     </GridBackground>
 
@@ -49,7 +48,5 @@ import { portofolioItems } from '@/lib/allData';
         />
     </section>
 
-    <!-- Hapus HTML lama dan panggil komponen CTA animasi -->
     <AnimatedCTA client:visible />
-
 </MainLayout>
